Don't strip message field from container logs
Some logs do not have a `log` field, and only have a `message` field. So
to err on the side of not losing any logs we keep the `message` field
even though it may sometimes duplicate the `log` field. I'm still not
certain when/why this occurs.

Add the `log` field of the record to an empty string if the field is
missing to avoid errors in the following parser block caused by the
missing field. At some point in the future when everything is running
containerd and therefore uses the `message` field, we will have to
update this workaround for `message`, and switch the parser block to use
the `message` field.