fix: drop chunks that fail elasticsearch submission earlier
The logs-concentrator has a buffer to retry failed log submission to
elasticsearch. This is useful for situations when elasticsearch is
temporarily unavailable.

However occasionally there are problems with submission to elasticearch
which is due to the data itself such as:

400 - Rejected by Elasticsearch [error type]: mapper_parsing_exception [reason]: 'Could not dynamically add mapping for field [cluster.name]. Existing mapping for [cluster] must be of type object but found [text].'

This chunk is never going to successfully submit to elasticsearch and
should just be dropped. We break up chunks by index_name, so only the
project with the badly shaped logs will have logs dropped.

By default fluentd will retry chunk submission for 72h which is far too
long. This change reduces that period to 12h which hopefully is long
enough to deal with any elasticsearch outages, but which also means we
can drop chunks which will never submit more quickly to help avoid
overflowing buffers.