feat: add feature to post-process router logs
This feature allows injecting snippets of fluentd configuration into the
router logs processing pipeline directly before the logs are sent to the
output.

This feature is designed to allow injecting logic to manipulate or add
fields to router log records which make the log records easier to use in
Opensearch Dashboards. For example this can be used to inject a boolean
field into log records which can be used instead of a slow Opensearch
regex search.